PyNet-main/requirements.txt
pyopenssl
pycryptodome

requests~=2.31.0
flask~=3.0.0